Moved server implementation to CPP file and added missing thread join.
